+++ src/sdk/main.js
@@ -24,6 +24,7 @@ import {delay, flush, destroy as schedulerDestroy} from './scheduler'
 import event from './event'
 import sdkClick from './sdk-click'
 import ActivityState from './activity-state'
+import { STORAGE_TYPES } from './constants'
 
 type InitConfigT = $ReadOnly<{|...InitOptionsT, ...LogOptionsT|}>
 
@@ -34,6 +35,14 @@ type InitConfigT = $ReadOnly<{|...InitOptionsT, ...LogOptionsT|}>
  * @private
  */
 let _options: ?InitOptionsT = null
+
+/**
+ * Flag to mark id sdk is in starting process
+ *
+ * @type {boolean}
+ * @private
+ */
+let _isInitialising: boolean = false
 
 /**
  * Flag to mark if sdk is started
@@ -59,17 +68,9 @@ let _isInstalled: boolean = false
  * @param {string} logOutput
  */
 function initSdk ({logLevel, logOutput, ...options}: InitConfigT = {}): void {
-
   Logger.setLogLevel(logLevel, logOutput)
 
-  if (!Storage) {
-    Logger.error('Adjust SDK can not start, there is no storage available')
-    return
-  }
-
-  Logger.info(`Available storage is ${Storage.type}`)
-
-  if (Config.isInitialised()) {
+  if (_isInitialised()) {
     Logger.error('You already initiated your instance')
     return
   }
@@ -78,9 +79,22 @@ function initSdk ({logLevel, logOutput, ...options}: InitConfigT = {}): void {
     return
   }
 
-  _options = {...options}
+  _isInitialising = true
 
-  _start(options)
+  Storage.init(options.namespace)
+    .then(availableStorage => {
+
+      if (availableStorage.type === STORAGE_TYPES.NO_STORAGE) {
+        Logger.error('Adjust SDK can not start, there is no storage available')
+        return
+      }
+
+      Logger.info(`Available storage is ${availableStorage.type}`)
+
+      _options = { ...options }
+
+      _start(options)
+    })
 }
 
 /**
@@ -244,6 +258,15 @@ function _handleGdprForgetMe (): void {
 }
 
 /**
+ * Check if sdk initialisation was started
+ *
+ * @private
+ */
+function _isInitialised (): boolean {
+  return _isInitialising || Config.isInitialised()
+}
+
+/**
  * Pause sdk by canceling:
  * - queue execution
  * - session watch
@@ -252,6 +275,7 @@ function _handleGdprForgetMe (): void {
  * @private
  */
 function _pause (): void {
+  _isInitialising = false
   _isStarted = false
 
   schedulerDestroy()
@@ -284,6 +308,8 @@ function _shutdown (async): void {
  * @private
  */
 function _destroy (): void {
+  _isInstalled = false
+
   _shutdown()
   gdprForgetDestroy()
 
@@ -331,6 +357,7 @@ function _continue (activityState: ActivityStateMapT): Promise<void> {
 
   return sessionWatch()
     .then(() => {
+      _isInitialising = false
       _isStarted = true
 
       if (isInstalled) {
@@ -345,7 +372,7 @@ function _continue (activityState: ActivityStateMapT): Promise<void> {
  */
 function _handleSdkInstalled () {
   _isInstalled = true
-  
+
   flush()
 
   unsubscribe('sdk:installed')
@@ -427,7 +454,7 @@ function _start (options: InitOptionsT): void {
  * @private
  */
 function _preCheck (description: string, callback: () => mixed, {schedule, stopBeforeInit}: {schedule?: boolean, stopBeforeInit?: boolean} = {}) {
-  if (!Storage) {
+  if (Storage.getType() === STORAGE_TYPES.NO_STORAGE) {
     Logger.log(`Adjust SDK can not ${description}, no storage available`)
     return
   }
@@ -437,19 +464,23 @@ function _preCheck (description: string, callback: () => mixed, {schedule, stopB
     return
   }
 
-  if (schedule && stopBeforeInit && !Config.isInitialised()) {
+  if (schedule && stopBeforeInit && !_isInitialised()) {
     Logger.error(`Adjust SDK can not ${description}, sdk instance is not initialized`)
     return
   }
 
   if (typeof callback === 'function') {
-    if (schedule && !(_isInstalled && _isStarted) && (stopBeforeInit || Config.isInitialised())) {
+    if (schedule && !(_isInstalled && _isStarted) && (stopBeforeInit || _isInitialised())) {
       delay(callback, description)
       Logger.log(`Running ${description} is delayed until Adjust SDK is up`)
     } else {
       callback()
     }
   }
+}
+
+function _clearDatabase () {
+  return Storage.deleteDatabase()
 }
 
 const Adjust = {
@@ -468,7 +499,8 @@ const Adjust = {
   gdprForgetMe,
   disableThirdPartySharing,
   __testonly__: {
-    destroy: _destroy
+    destroy: _destroy,
+    clearDatabase: _clearDatabase
   }
 }