+++ src/sdk/types.js
@@ -160,8 +160,10 @@ export type BaseParamsT = $ReadOnly<$Shape<{
 }>>
 
 export type CustomConfigT = $ReadOnly<$Shape<{
+  urlStrategy: 'india' | 'china',
   customUrl: string,
-  eventDeduplicationListLimit: number
+  eventDeduplicationListLimit: number,
+  namespace: string
 }>>
 
 export type LogOptionsT = $ReadOnly<$Shape<{|
@@ -174,8 +176,10 @@ export type InitOptionsT = $ReadOnly<$Shape<{|
   environment: $PropertyType<BaseParamsT, 'environment'>,
   defaultTracker: $PropertyType<BaseParamsT, 'defaultTracker'>,
   externalDeviceId: $PropertyType<BaseParamsT, 'externalDeviceId'>,
+  urlStrategy: $PropertyType<CustomConfigT, 'urlStrategy'>,
   customUrl: $PropertyType<CustomConfigT, 'customUrl'>,
   eventDeduplicationListLimit: $PropertyType<CustomConfigT, 'eventDeduplicationListLimit'>,
+  namespace: $PropertyType<CustomConfigT, 'namespace'>,
   attributionCallback: (string, Object) => mixed
 |}>>